Move delete account endpoint to account
